rework db query limits: introduce new class service.QueryLimiter
