+++ app/service/common.go
@@ -20,25 +20,3 @@ type Base struct {
 func (srv *Base) GetUser(r *http.Request) *database.User {
 	return auth.UserFromContext(r.Context(), srv.Store.Users())
 }
-
-// SetQueryLimit limits the number of records of the given query according to the `limit` request parameter
-// The default limit is 500
-// The optional `limits` argument can be either [default_limit] or [default_limit, max_possible_limit]
-func SetQueryLimit(r *http.Request, db *database.DB, limits ...int) *database.DB {
-	limit, err := ResolveURLQueryGetInt64Field(r, "limit")
-	defaultLimitToUse := int64(500)
-	maxPossibleLimit := int64(1000)
-	if len(limits) > 0 {
-		defaultLimitToUse = int64(limits[0])
-	}
-	if len(limits) > 1 {
-		maxPossibleLimit = int64(limits[1])
-	}
-	if err != nil || limit < 0 {
-		limit = defaultLimitToUse
-	}
-	if limit > maxPossibleLimit {
-		limit = maxPossibleLimit
-	}
-	return db.Limit(limit)
-}
